refactor(tests): move large inline test modules to sibling files
Six files carried 200+ line #[cfg(test)] blocks inline while ~40 other modules
already used <module>/tests.rs, so which convention a file follows could only be
learned by opening it. Moves the nine blocks over the line and writes the
threshold down in AGENTS.md; smaller blocks stay inline on purpose.

Pure relocation: module names are unchanged, so cargo test filters still
address the same sets, and 'cargo test --lib -- --list' is identical before
and after (2273 tests, same paths).
