+++ src/target_parser/mod.rs
@@ -479,567 +479,7 @@ pub fn parse_raw_http_request(raw: &str) -> Result<Target, Box<dyn std::error::E
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    // The client cache is process-global and shared with every other test in
-    // this binary that builds a Client (e.g. the WAF tests). Asserting on the
-    // cache's *total* length is therefore racy: a concurrent test can insert
-    // an entry between two measurements, which previously surfaced as a
-    // `same key must reuse: left 1, right 2` flake. Each test below instead
-    // scopes its assertions to a timeout value no other test uses, so foreign
-    // inserts (which carry different timeouts) can't perturb the count. This
-    // makes the cache tests safe to run concurrently with no shared lock.
-    //
-    // Timeouts reserved for these tests' cache keys. The isolation guarantee
-    // rests on no other test building a Client with one of these values, so
-    // keep them unique to this module and do not reuse them elsewhere.
-    const REUSE_TIMEOUT: u64 = 60_001;
-    const DISTINCT_TIMEOUT_A: u64 = 60_002;
-    const DISTINCT_TIMEOUT_B: u64 = 60_003;
-    const INSECURE_TIMEOUT: u64 = 60_004;
-
-    /// Count cached Clients whose key carries `timeout`. Scoping by a
-    /// per-test-unique timeout isolates the measurement from any other test
-    /// that builds a Client into the same process-global cache.
-    fn cache_entries_with_timeout(timeout: u64) -> usize {
-        client_cache()
-            .lock()
-            .map(|g| g.keys().filter(|(t, _, _, _)| *t == timeout).count())
-            .unwrap_or(0)
-    }
-
-    /// Building twice with the same key reuses the cached Client instead of
-    /// creating a second entry.
-    #[test]
-    fn test_client_cache_reuses_for_same_key() {
-        let mut t = parse_target("http://example.com").unwrap();
-        t.timeout = REUSE_TIMEOUT;
-        t.follow_redirects = false;
-        t.proxy = None;
-        let _ = t.build_client().unwrap();
-        let after_first = cache_entries_with_timeout(t.timeout);
-        let _ = t.build_client().unwrap();
-        let after_second = cache_entries_with_timeout(t.timeout);
-        assert_eq!(after_first, after_second, "same key must reuse");
-        assert_eq!(after_first, 1, "exactly one entry for the shared key");
-    }
-
-    /// Two targets differing only in an input that affects Client
-    /// construction (here, timeout) occupy two distinct cache entries.
-    #[test]
-    fn test_client_cache_separates_distinct_keys() {
-        let mut a = parse_target("http://example.com").unwrap();
-        a.timeout = DISTINCT_TIMEOUT_A;
-        let mut b = parse_target("http://example.com").unwrap();
-        b.timeout = DISTINCT_TIMEOUT_B; // distinct key
-        let _ = a.build_client().unwrap();
-        let _ = b.build_client().unwrap();
-        let _ = a.build_client().unwrap();
-        assert_eq!(cache_entries_with_timeout(a.timeout), 1);
-        assert_eq!(cache_entries_with_timeout(b.timeout), 1);
-    }
-
-    /// `insecure` is part of the Client cache key, so toggling TLS
-    /// verification yields two distinct cached Clients rather than silently
-    /// reusing one built with the other posture.
-    #[test]
-    fn test_client_cache_separates_on_insecure() {
-        let mut secure = parse_target("https://example.com").unwrap();
-        secure.timeout = INSECURE_TIMEOUT;
-        secure.insecure = false;
-        let mut insecure = parse_target("https://example.com").unwrap();
-        insecure.timeout = INSECURE_TIMEOUT;
-        insecure.insecure = true;
-        let _ = secure.build_client().unwrap();
-        let _ = insecure.build_client().unwrap();
-        // Rebuild to confirm reuse (no third entry created).
-        let _ = secure.build_client().unwrap();
-        assert_eq!(
-            cache_entries_with_timeout(INSECURE_TIMEOUT),
-            2,
-            "secure and insecure clients must not share a cache entry"
-        );
-    }
-
-    /// Scanner default: a freshly parsed target trusts invalid certs unless
-    /// the caller opts into validation.
-    #[test]
-    fn test_parse_target_defaults_to_insecure() {
-        let target = parse_target("https://example.com").unwrap();
-        assert!(
-            target.insecure,
-            "insecure must default to true (scanner mode)"
-        );
-    }
-
-    #[test]
-    fn test_parse_target_with_scheme() {
-        let target = parse_target("https://example.com").unwrap();
-        assert_eq!(target.url.as_str(), "https://example.com/");
-        assert_eq!(target.method, "GET");
-        assert!(target.data.is_none());
-        assert!(target.headers.is_empty());
-        assert!(target.cookies.is_empty());
-        assert!(target.user_agent.is_none());
-        assert!(target.reflection_params.is_empty());
-        assert_eq!(target.timeout, 10);
-        assert_eq!(target.delay, 0);
-        assert!(target.proxy.is_none());
-        assert_eq!(target.workers, 10);
-        assert!(!target.follow_redirects);
-    }
-
-    #[test]
-    fn test_parse_target_without_scheme() {
-        let target = parse_target("example.com").unwrap();
-        assert_eq!(target.url.as_str(), "http://example.com/");
-        assert_eq!(target.method, "GET");
-    }
-
-    #[test]
-    fn test_parse_target_invalid_url() {
-        assert!(parse_target("invalid url").is_err());
-    }
-
-    /// Explicit non-http(s) authority-form schemes are rejected with an
-    /// actionable error instead of being silently mangled into a bogus host
-    /// (`ftp://127.0.0.1/x` used to become `http://ftp//127.0.0.1/x` and fail
-    /// DNS as if the user's host were down).
-    #[test]
-    fn test_parse_target_rejects_non_http_scheme() {
-        for bad in [
-            "ftp://127.0.0.1/x?q=1",
-            "file:///etc/passwd",
-            "gopher://127.0.0.1/x",
-            "ws://127.0.0.1/x",
-            "FTP://127.0.0.1/x", // case-insensitive
-        ] {
-            let err = parse_target(bad).expect_err("non-http(s) scheme must be rejected");
-            assert!(
-                err.to_string().contains("unsupported URL scheme"),
-                "expected an unsupported-scheme error for {bad}, got: {err}"
-            );
-        }
-    }
-
-    /// Scheme-less inputs that merely *contain* a colon (or a `://` inside the
-    /// query) must keep working — the rejection is anchored on a contiguous
-    /// leading `scheme://`, so these still get the `http://` prefix.
-    #[test]
-    fn test_parse_target_keeps_schemeless_inputs() {
-        for ok in [
-            "127.0.0.1:8771/ctx/vuln_body?q=1",        // host:port
-            "user:pass@127.0.0.1:8771/x",              // userinfo colon
-            "example.com/p?next=https://evil.com&q=1", // `://` only in query
-            "example.com",
-        ] {
-            let t = parse_target(ok).unwrap_or_else(|e| panic!("{ok} should parse, got: {e}"));
-            assert!(
-                t.url.as_str().starts_with("http://"),
-                "{ok} should be prefixed with http://, got {}",
-                t.url.as_str()
-            );
-        }
-    }
-
-    #[test]
-    fn test_parse_target_with_path() {
-        let target = parse_target("https://example.com/path/to/resource").unwrap();
-        assert_eq!(target.url.as_str(), "https://example.com/path/to/resource");
-        assert_eq!(target.method, "GET");
-    }
-
-    #[test]
-    fn test_parse_target_with_query() {
-        let target = parse_target("https://example.com?param=value").unwrap();
-        assert_eq!(target.url.as_str(), "https://example.com/?param=value");
-        assert_eq!(target.method, "GET");
-    }
-
-    #[test]
-    fn test_parse_target_with_fragment() {
-        let target = parse_target("https://example.com#section").unwrap();
-        assert_eq!(target.url.as_str(), "https://example.com/#section");
-        assert_eq!(target.method, "GET");
-    }
-
-    #[test]
-    fn test_parse_target_with_port() {
-        let target = parse_target("https://example.com:8080").unwrap();
-        assert_eq!(target.url.as_str(), "https://example.com:8080/");
-        assert_eq!(target.method, "GET");
-    }
-
-    #[test]
-    fn test_parse_target_ip_address() {
-        let target = parse_target("http://192.168.1.1").unwrap();
-        assert_eq!(target.url.as_str(), "http://192.168.1.1/");
-        assert_eq!(target.method, "GET");
-    }
-
-    #[test]
-    fn test_parse_target_localhost() {
-        let target = parse_target("localhost:3000").unwrap();
-        assert_eq!(target.url.as_str(), "http://localhost:3000/");
-        assert_eq!(target.method, "GET");
-    }
-
-    #[test]
-    fn test_parse_target_unicode_domain() {
-        // URL parsing converts unicode domains to punycode
-        let target = parse_target("https://例え.テスト").unwrap();
-        assert!(target.url.as_str().contains("xn--r8jz45g.xn--zckzah"));
-        assert_eq!(target.method, "GET");
-    }
-
-    #[test]
-    fn test_parse_target_empty_string() {
-        assert!(parse_target("").is_err());
-    }
-
-    #[test]
-    fn test_parse_target_only_spaces() {
-        assert!(parse_target("   ").is_err());
-    }
-
-    #[test]
-    fn test_parse_target_invalid_scheme() {
-        assert!(parse_target("://example.com").is_err());
-    }
-
-    #[test]
-    fn test_parse_method_url_body_post_with_body() {
-        let (method, url, body) = parse_method_url_body("POST https://example.com/test a=b");
-        assert_eq!(method, "POST");
-        assert_eq!(url, "https://example.com/test");
-        assert_eq!(body, Some("a=b".to_string()));
-    }
-
-    #[test]
-    fn test_parse_method_url_body_get_without_body() {
-        let (method, url, body) = parse_method_url_body("GET https://example.com/path");
-        assert_eq!(method, "GET");
-        assert_eq!(url, "https://example.com/path");
-        assert_eq!(body, None);
-    }
-
-    #[test]
-    fn test_parse_method_url_body_put_with_json() {
-        let (method, url, body) =
-            parse_method_url_body("PUT https://api.example.com {\"key\":\"value\"}");
-        assert_eq!(method, "PUT");
-        assert_eq!(url, "https://api.example.com");
-        assert_eq!(body, Some("{\"key\":\"value\"}".to_string()));
-    }
-
-    #[test]
-    fn test_parse_method_url_body_plain_url() {
-        let (method, url, body) = parse_method_url_body("https://example.com");
-        assert_eq!(method, "GET");
-        assert_eq!(url, "https://example.com");
-        assert_eq!(body, None);
-    }
-
-    #[test]
-    fn test_parse_method_url_body_lowercase_method() {
-        let (method, url, body) = parse_method_url_body("post https://example.com data=test");
-        assert_eq!(method, "POST");
-        assert_eq!(url, "https://example.com");
-        assert_eq!(body, Some("data=test".to_string()));
-    }
-
-    #[test]
-    fn test_parse_method_url_body_delete() {
-        let (method, url, body) =
-            parse_method_url_body("DELETE https://api.example.com/resource/123");
-        assert_eq!(method, "DELETE");
-        assert_eq!(url, "https://api.example.com/resource/123");
-        assert_eq!(body, None);
-    }
-
-    #[test]
-    fn test_parse_method_url_body_options() {
-        let (method, url, body) = parse_method_url_body("OPTIONS https://example.com/api");
-        assert_eq!(method, "OPTIONS");
-        assert_eq!(url, "https://example.com/api");
-        assert_eq!(body, None);
-    }
-
-    #[test]
-    fn test_parse_method_url_body_query_with_json_body() {
-        // RFC 10008 QUERY — safe/idempotent method with a body.
-        let (method, url, body) =
-            parse_method_url_body("QUERY https://example.com/search {\"q\":\"test\"}");
-        assert_eq!(method, "QUERY");
-        assert_eq!(url, "https://example.com/search");
-        assert_eq!(body, Some("{\"q\":\"test\"}".to_string()));
-    }
-
-    #[test]
-    fn test_parse_method_url_body_query_lowercase() {
-        let (method, url, body) = parse_method_url_body("query https://example.com/search a=b");
-        assert_eq!(method, "QUERY");
-        assert_eq!(url, "https://example.com/search");
-        assert_eq!(body, Some("a=b".to_string()));
-    }
-
-    #[test]
-    fn test_is_raw_http_request_query() {
-        assert!(is_raw_http_request(
-            "QUERY /search HTTP/1.1\r\nHost: example.com\r\n\r\n{\"q\":\"test\"}"
-        ));
-        assert!(!is_raw_http_request("QUERY https://example.com/search"));
-    }
-
-    #[test]
-    fn test_parse_target_with_method_query() {
-        let target =
-            parse_target_with_method("QUERY https://example.com/search {\"q\":\"x\"}").unwrap();
-        assert_eq!(target.method, "QUERY");
-        assert_eq!(target.url.as_str(), "https://example.com/search");
-        assert_eq!(target.data, Some("{\"q\":\"x\"}".to_string()));
-    }
-
-    #[test]
-    fn test_parse_target_with_method_post() {
-        let target = parse_target_with_method("POST https://www.hahwul.com/post-test a=b").unwrap();
-        assert_eq!(target.method, "POST");
-        assert_eq!(target.url.as_str(), "https://www.hahwul.com/post-test");
-        assert_eq!(target.data, Some("a=b".to_string()));
-    }
-
-    #[test]
-    fn test_parse_target_with_method_get() {
-        let target = parse_target_with_method("GET https://example.com/path").unwrap();
-        assert_eq!(target.method, "GET");
-        assert_eq!(target.url.as_str(), "https://example.com/path");
-        assert_eq!(target.data, None);
-    }
-
-    #[test]
-    fn test_parse_target_with_method_plain_url() {
-        let target = parse_target_with_method("https://example.com").unwrap();
-        assert_eq!(target.method, "GET");
-        assert_eq!(target.url.as_str(), "https://example.com/");
-        assert_eq!(target.data, None);
-    }
-
-    #[test]
-    fn test_parse_target_with_method_body_with_spaces() {
-        let target =
-            parse_target_with_method("POST https://example.com/api name=John Doe").unwrap();
-        assert_eq!(target.method, "POST");
-        assert_eq!(target.url.as_str(), "https://example.com/api");
-        assert_eq!(target.data, Some("name=John Doe".to_string()));
-    }
-}
+mod tests;
 
 #[cfg(test)]
-mod raw_http_tests {
-    use super::*;
-
-    #[test]
-    fn test_is_raw_http_request_true() {
-        let raw = "GET /path HTTP/1.1\r\nHost: example.com\r\n\r\n";
-        assert!(is_raw_http_request(raw));
-    }
-
-    #[test]
-    fn test_is_raw_http_request_false() {
-        let not_raw = "https://example.com/path";
-        assert!(!is_raw_http_request(not_raw));
-    }
-
-    #[test]
-    fn test_parse_raw_http_absolute_form() {
-        let raw = "GET http://example.com/level1/frame HTTP/1.1\r\nUser-Agent: Dalfox\r\nCookie: sid=abc; a=b\r\n\r\n";
-        let t = parse_raw_http_request(raw).expect("should parse absolute-form request");
-        assert_eq!(t.method, "GET");
-        assert_eq!(t.url.as_str(), "http://example.com/level1/frame");
-        assert_eq!(t.user_agent.as_deref(), Some("Dalfox"));
-        assert!(t.cookies.iter().any(|(k, v)| k == "sid" && v == "abc"));
-        assert!(t.cookies.iter().any(|(k, v)| k == "a" && v == "b"));
-    }
-
-    #[test]
-    fn test_parse_raw_http_skips_empty_name_cookies() {
-        // A `=value` (empty-name) segment in a raw-request Cookie header must
-        // not produce an empty-name cookie pair; the well-formed pairs survive.
-        // Mirrors har.rs::skips_empty_name_cookies_from_header so the twin
-        // guards have symmetric coverage.
-        let raw = "GET http://example.com/ HTTP/1.1\r\nCookie: a=1; =orphan; b=2\r\n\r\n";
-        let t = parse_raw_http_request(raw).expect("should parse");
-        assert_eq!(t.cookies.len(), 2);
-        assert!(t.cookies.iter().any(|(k, v)| k == "a" && v == "1"));
-        assert!(t.cookies.iter().any(|(k, v)| k == "b" && v == "2"));
-        assert!(
-            t.cookies.iter().all(|(k, _)| !k.is_empty()),
-            "no empty-name cookie should be retained"
-        );
-    }
-
-    #[test]
-    fn test_parse_raw_http_origin_form() {
-        let raw = "GET /level1/frame HTTP/1.1\r\nHost: vulnerable.com\r\n\r\n";
-        let t = parse_raw_http_request(raw).expect("should parse origin-form with Host");
-        assert_eq!(t.url.as_str(), "http://vulnerable.com/level1/frame");
-        assert_eq!(t.method, "GET");
-    }
-
-    #[test]
-    fn test_parse_raw_http_https_host_port() {
-        let raw = "GET /p HTTP/1.1\r\nHost: secure.example.com:443\r\n\r\n";
-        let t = parse_raw_http_request(raw).expect("should infer https from :443");
-        let u = &t.url;
-        assert_eq!(u.scheme(), "https");
-        assert_eq!(u.host_str(), Some("secure.example.com"));
-        assert_eq!(u.port_or_known_default(), Some(443));
-    }
-
-    #[test]
-    fn test_parse_raw_http_missing_host_errors() {
-        let raw = "GET /only-path HTTP/1.1\r\nUser-Agent: X\r\n\r\n";
-        assert!(parse_raw_http_request(raw).is_err());
-    }
-
-    #[test]
-    fn test_parse_raw_http_with_body() {
-        let raw = "POST /submit HTTP/1.1\r\nHost: example.com\r\nContent-Type: application/x-www-form-urlencoded\r\n\r\na=1&b=2";
-        let t = parse_raw_http_request(raw).expect("should parse with body");
-        assert_eq!(t.method, "POST");
-        assert_eq!(t.url.as_str(), "http://example.com/submit");
-        assert_eq!(t.data.as_deref(), Some("a=1&b=2"));
-    }
-
-    #[test]
-    fn raw_http_body_preserves_crlf_line_endings() {
-        // A multipart body's CRLFs are load-bearing: RFC 7578 parsers match
-        // `\r\n--boundary`. Rebuilding the body from `lines()` joined with `\n`
-        // turned every one of them into a bare LF, so a strict server saw zero
-        // parts and the whole body-parameter scan silently tested nothing.
-        //
-        // The body's *final* terminator is a separate question — it is a file
-        // artifact and is dropped (see
-        // `raw_http_body_drops_exactly_one_trailing_newline`); the close
-        // delimiter itself and every internal CRLF survive.
-        let body = "--X\r\nContent-Disposition: form-data; name=\"q\"\r\n\r\nv\r\n--X--";
-        let raw = format!(
-            "POST /u HTTP/1.1\r\nHost: e.com\r\nContent-Type: multipart/form-data; boundary=X\r\n\r\n{body}\r\n"
-        );
-        let t = parse_raw_http_request(&raw).expect("should parse");
-        assert_eq!(t.data.as_deref(), Some(body));
-    }
-
-    #[test]
-    fn raw_http_body_preserves_a_leading_blank_line() {
-        // The old `lines()` fold started the accumulator empty, so a body that
-        // legitimately begins with a blank line lost it.
-        let raw = "POST /u HTTP/1.1\r\nHost: e.com\r\n\r\n\r\nreal body";
-        let t = parse_raw_http_request(raw).expect("should parse");
-        assert_eq!(t.data.as_deref(), Some("\r\nreal body"));
-    }
-
-    #[test]
-    fn raw_http_lf_only_request_still_finds_its_body() {
-        let raw = "POST /u HTTP/1.1\nHost: e.com\n\na=1&b=2";
-        let t = parse_raw_http_request(raw).expect("should parse");
-        assert_eq!(t.data.as_deref(), Some("a=1&b=2"));
-    }
-
-    #[test]
-    fn raw_http_without_a_body_has_none() {
-        let raw = "GET /u HTTP/1.1\r\nHost: e.com\r\n\r\n";
-        let t = parse_raw_http_request(raw).expect("should parse");
-        assert_eq!(t.data, None);
-    }
-
-    #[test]
-    fn raw_http_body_drops_exactly_one_trailing_newline() {
-        // Every text editor terminates a file with a newline; keeping it made
-        // the last body parameter's value `2\n`, which then rode along
-        // percent-encoded on every request of the scan.
-        let t = parse_raw_http_request("POST /u HTTP/1.1\r\nHost: e\r\n\r\na=1&b=2\r\n")
-            .expect("should parse");
-        assert_eq!(t.data.as_deref(), Some("a=1&b=2"));
-        let lf =
-            parse_raw_http_request("POST /u HTTP/1.1\nHost: e\n\na=1&b=2\n").expect("should parse");
-        assert_eq!(lf.data.as_deref(), Some("a=1&b=2"));
-        // Only ONE terminator goes: a body that really ends in a blank line
-        // keeps the rest.
-        let two = parse_raw_http_request("POST /u HTTP/1.1\r\nHost: e\r\n\r\na=1\r\n\r\n")
-            .expect("should parse");
-        assert_eq!(two.data.as_deref(), Some("a=1\r\n"));
-    }
-
-    #[test]
-    fn raw_http_body_survives_a_whitespace_bearing_separator_line() {
-        // The header loop ends on the first line whose `trim_end()` is empty,
-        // so a separator carrying a stray space ends the headers there.
-        // Scanning for a literal `\r\n\r\n` missed it and reported NO body,
-        // silently dropping every body parameter from the scan.
-        let t = parse_raw_http_request("POST /u HTTP/1.1\r\nHost: e\r\n \r\na=1&b=2")
-            .expect("should parse");
-        assert_eq!(t.data.as_deref(), Some("a=1&b=2"));
-    }
-
-    #[test]
-    fn raw_http_multipart_close_delimiter_is_preserved() {
-        // The trailing-newline strip must not eat a multipart body's own CRLF
-        // structure; only the final terminator goes.
-        let body = "--X\r\nContent-Disposition: form-data; name=\"q\"\r\n\r\nv\r\n--X--";
-        let raw = format!(
-            "POST /u HTTP/1.1\r\nHost: e\r\nContent-Type: multipart/form-data; boundary=X\r\n\r\n{body}\r\n"
-        );
-        let t = parse_raw_http_request(&raw).expect("should parse");
-        assert_eq!(t.data.as_deref(), Some(body));
-    }
-
-    #[test]
-    fn raw_http_strips_stale_framing_and_accept_encoding_headers() {
-        // A stale Content-Length / Transfer-Encoding must NOT be forwarded:
-        // reqwest recomputes them and a stale value truncates the
-        // payload-injected body (body-param XSS silently missed). Accept-Encoding
-        // must be dropped so reqwest's transparent decompression stays on.
-        let raw = "POST /submit HTTP/1.1\r\nHost: example.com\r\nContent-Length: 7\r\nTransfer-Encoding: chunked\r\nAccept-Encoding: gzip\r\nConnection: keep-alive\r\nContent-Type: application/x-www-form-urlencoded\r\n\r\na=1&b=2";
-        let t = parse_raw_http_request(raw).expect("should parse");
-        for stripped in [
-            "content-length",
-            "transfer-encoding",
-            "accept-encoding",
-            "connection",
-        ] {
-            assert!(
-                !t.headers
-                    .iter()
-                    .any(|(k, _)| k.eq_ignore_ascii_case(stripped)),
-                "{stripped} must be stripped, got {:?}",
-                t.headers
-            );
-        }
-        // A normal header survives verbatim.
-        assert!(
-            t.headers
-                .iter()
-                .any(|(k, _)| k.eq_ignore_ascii_case("content-type"))
-        );
-    }
-
-    #[test]
-    fn raw_http_does_not_retain_original_cookie_header() {
-        // The Cookie header is split into `cookies` only; keeping it in `headers`
-        // too made per-cookie probing emit both the original and the mutated
-        // Cookie (reqwest appends), so the payload could fail to land.
-        let raw = "GET /p HTTP/1.1\r\nHost: example.com\r\nCookie: sid=abc; a=b\r\n\r\n";
-        let t = parse_raw_http_request(raw).expect("should parse");
-        assert!(
-            !t.headers
-                .iter()
-                .any(|(k, _)| k.eq_ignore_ascii_case("cookie")),
-            "original Cookie header must not be retained in headers, got {:?}",
-            t.headers
-        );
-        assert!(t.cookies.iter().any(|(k, v)| k == "sid" && v == "abc"));
-        assert!(t.cookies.iter().any(|(k, v)| k == "a" && v == "b"));
-    }
-}
+mod raw_http_tests;
